Fold agent_addr into find/discover, rename addresses.py, expose address properties
Continues the address-parsing refactor:

- Rename api/addresses.py to api/address_parsers.py to better reflect that
  it only contains parsers (the address types themselves live in primitives.py).
- Delete api/agent_addr.py: with typed addresses as the norm at the API
  boundary, there is no longer a reason to segregate address-accepting
  functions. discover_by_address moved to api/discover.py;
  find_agent_by_address and find_agents_by_addresses moved to api/find.py.
  Helpers that are no longer called externally are now private.
- Make host_addresses_match a method on HostAddress (HostAddress.matches).
- Move collect_required_provider_names from api/address_parsers.py into
  api/find.py (now private; was only used by find logic).
- Add cached `address` properties to AgentDetails and HostDetails so callers
  can pass typed addresses directly to api functions instead of reconstructing
  them from id/name/provider fields.
