+++ libs/mngr/imbue/mngr/api/exec.py
@@ -11,11 +11,10 @@ from imbue.imbue_common.frozen_model import FrozenModel
 from imbue.imbue_common.logging import log_call
 from imbue.imbue_common.logging import log_span
 from imbue.imbue_common.mutable_model import MutableModel
-from imbue.mngr.api.addresses import AgentAddress
-from imbue.mngr.api.agent_addr import find_agent_by_address
-from imbue.mngr.api.agent_addr import find_agents_by_addresses
 from imbue.mngr.api.find import AgentMatch
 from imbue.mngr.api.find import ensure_host_started
+from imbue.mngr.api.find import find_agent_by_address
+from imbue.mngr.api.find import find_agents_by_addresses
 from imbue.mngr.api.find import group_agents_by_host
 from imbue.mngr.api.providers import get_provider_instance
 from imbue.mngr.config.data_types import MngrContext
@@ -24,6 +23,7 @@ from imbue.mngr.errors import MngrError
 from imbue.mngr.errors import UserInputError
 from imbue.mngr.interfaces.agent import AgentInterface
 from imbue.mngr.interfaces.host import OnlineHostInterface
+from imbue.mngr.primitives import AgentAddress
 from imbue.mngr.primitives import AgentId
 from imbue.mngr.primitives import AgentName
 from imbue.mngr.primitives import ErrorBehavior
