+++ libs/mngr/imbue/mngr/cli/exec.py
@@ -5,7 +5,6 @@ import click
 from click_option_group import optgroup
 from loguru import logger
 
-from imbue.mngr.api.addresses import AgentAddress
 from imbue.mngr.api.exec import ExecResult
 from imbue.mngr.api.exec import MissingOuterBehavior
 from imbue.mngr.api.exec import MultiExecResult
@@ -30,6 +29,7 @@ from imbue.mngr.cli.stdin_utils import expand_stdin_placeholder
 from imbue.mngr.config.data_types import CommonCliOptions
 from imbue.mngr.config.data_types import OutputOptions
 from imbue.mngr.errors import UserInputError
+from imbue.mngr.primitives import AgentAddress
 from imbue.mngr.primitives import ErrorBehavior
 from imbue.mngr.primitives import OutputFormat
