+++ libs/mngr/imbue/mngr/cli/provision_test.py
@@ -6,11 +6,11 @@ import pluggy
 import pytest
 from click.testing import CliRunner
 
-from imbue.mngr.api.addresses import AgentAddress
 from imbue.mngr.cli.provision import ProvisionCliOptions
 from imbue.mngr.cli.provision import _output_result
 from imbue.mngr.cli.provision import provision
 from imbue.mngr.config.data_types import OutputOptions
+from imbue.mngr.primitives import AgentAddress
 from imbue.mngr.primitives import AgentName
 from imbue.mngr.primitives import OutputFormat
