+++ libs/mngr/imbue/mngr/cli/pull.py
@@ -4,9 +4,6 @@ import click
 from click_option_group import optgroup
 from loguru import logger
 
-from imbue.mngr.api.addresses import AgentAddress
-from imbue.mngr.api.addresses import HostAddress
-from imbue.mngr.api.addresses import SourceLocation
 from imbue.mngr.api.pull import pull_files
 from imbue.mngr.api.pull import pull_git
 from imbue.mngr.cli.address_params import AGENT_ADDRESS
@@ -23,6 +20,9 @@ from imbue.mngr.cli.output_helpers import output_sync_files_result
 from imbue.mngr.cli.output_helpers import output_sync_git_result
 from imbue.mngr.config.data_types import CommonCliOptions
 from imbue.mngr.errors import UserInputError
+from imbue.mngr.primitives import AgentAddress
+from imbue.mngr.primitives import HostAddress
+from imbue.mngr.primitives import SourceLocation
 from imbue.mngr.primitives import UncommittedChangesMode
