+++ libs/mngr/imbue/mngr/cli/start_test.py
@@ -6,11 +6,11 @@ import pluggy
 import pytest
 from click.testing import CliRunner
 
-from imbue.mngr.api.addresses import AgentAddress
 from imbue.mngr.cli.start import StartCliOptions
 from imbue.mngr.cli.start import _output_result
 from imbue.mngr.cli.start import start
 from imbue.mngr.config.data_types import OutputOptions
+from imbue.mngr.primitives import AgentAddress
 from imbue.mngr.primitives import AgentName
 from imbue.mngr.primitives import OutputFormat
