+++ libs/mngr/imbue/mngr/cli/stop_test.py
@@ -8,11 +8,11 @@ import pluggy
 import pytest
 from click.testing import CliRunner
 
-from imbue.mngr.api.addresses import AgentAddress
 from imbue.mngr.cli.stop import StopCliOptions
 from imbue.mngr.cli.stop import _output_result
 from imbue.mngr.cli.stop import stop
 from imbue.mngr.config.data_types import OutputOptions
+from imbue.mngr.primitives import AgentAddress
 from imbue.mngr.primitives import AgentName
 from imbue.mngr.primitives import OutputFormat
