+++ libs/mngr_schedule/imbue/mngr_schedule/implementations/modal/verification.py
@@ -20,7 +20,7 @@ from loguru import logger
 
 from imbue.concurrency_group.concurrency_group import ConcurrencyGroup
 from imbue.imbue_common.pure import pure
-from imbue.mngr.api.addresses import parse_agent_name_or_id
+from imbue.mngr.api.address_parsers import parse_agent_name_or_id
 from imbue.mngr.api.discover import discover_hosts_and_agents
 from imbue.mngr.api.find import find_and_maybe_start_agent
 from imbue.mngr.config.data_types import MngrContext
